+++ examples/entity_parsing_from_scratch.rs
@@ -28,7 +28,7 @@ fn main() {
         .build()
         .unwrap();
 
-    let sentence = "My favourite artists are the stones and the fab four";
+    let sentence = "My favourite artists are the stones and fab four";
     let extracted_entities = parser.run(sentence).unwrap();
     assert_eq!(extracted_entities,
                vec![
@@ -36,10 +36,12 @@ fn main() {
                        raw_value: "the stones".to_string(),
                        resolved_value: "The Rolling Stones".to_string(),
                        range: 25..35,
+                       matched_value: "the rolling stones".to_string()
                    },
                    ParsedValue {
-                       raw_value: "the fab four".to_string(),
+                       raw_value: "fab four".to_string(),
                        resolved_value: "The Beatles".to_string(),
-                       range: 40..52,
+                       range: 40..48,
+                       matched_value: "the fab four".to_string(),
                    }]);
 }