+++ CHANGELOG.md
@@ -1,19 +1,25 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.7.0] - 2019-04-16
+### Added
+- Add API to prepend entity values [#31](https://github.com/snipsco/gazetteer-entity-parser/pull/31)
+- Add matched value in API output [#32](https://github.com/snipsco/gazetteer-entity-parser/pull/32)
+
 ## [0.6.0] - 2018-11-09
 ### Changed
-- Optimize memory usage
-- Simpler pattern for errors
+- Optimize memory usage [#27](https://github.com/snipsco/gazetteer-entity-parser/pull/27)
+- Simpler pattern for errors [#27](https://github.com/snipsco/gazetteer-entity-parser/pull/27)
 
 ## [0.5.1] - 2018-10-15
 ### Changed
-- Fix bug affecting the backward expansion of possible matches starting with stop words
+- Fix bug affecting the backward expansion of possible matches starting with stop words [#25](https://github.com/snipsco/gazetteer-entity-parser/pull/25)
 
 ## [0.5.0] - 2018-10-01
 ### Changed
 - Clearer `ParserBuilder`'s API 
 
+[0.7.0]: https://github.com/snipsco/gazetteer-entity-parser/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/snipsco/gazetteer-entity-parser/compare/0.5.1...0.6.0
 [0.5.1]: https://github.com/snipsco/gazetteer-entity-parser/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/snipsco/gazetteer-entity-parser/compare/0.4.2...0.5.0